# Speed up token retrieval