@ankane 2021-04-23
Hey @cbillen, thanks for the PR! Agree `find_by` would be better.

fwiw, I believe they'll have similar performance if the database statistics are up-to-date. Can you try running `ANALYZE VERBOSE ahoy_visits;` and running the first query again to see if that improves it?