@cbillen 2021-04-23
I believe you are right
ANALYZE VERBOSE visits;

EXPLAIN ANALYZE
SELECT "visits".* FROM "visits" WHERE "visits"."visit_token" = '4208f558-2c45-4d9d-806b-b1bf14194665' ORDER BY "visits"."id" ASC LIMIT 1

Limit (cost=8.03..8.03 rows=1 width=658) (actual time=0.059..0.088 rows=1 loops=1)
--
-> Sort (cost=8.03..8.03 rows=1 width=658) (actual time=0.048..0.063 rows=1 loops=1)
Sort Key: id
Sort Method: quicksort Memory: 25kB
-> Index Scan using index_visits_on_visit_token on visits (cost=0.00..8.02 rows=1 width=658) (actual time=0.021..0.031 rows=1 loops=1)
Index Cond: ((visit_token)::text = '4208f558-2c45-4d9d-806b-b1bf14194665'::text)
Planning Time: 0.182 ms
Execution Time: 0.153 ms

But without the order by id, still faster by a smidge

Limit (cost=0.00..8.02 rows=1 width=658) (actual time=0.030..0.047 rows=1 loops=1)
--
-> Index Scan using index_visits_on_visit_token on visits (cost=0.00..8.02 rows=1 width=658) (actual time=0.020..0.024 rows=1 loops=1)
Index Cond: ((visit_token)::text = '4208f558-2c45-4d9d-806b-b1bf14194665'::text)
Planning Time: 0.092 ms
Execution Time: 0.106 ms


I have also experimented using hash index on postgresql, slightly faster too, another gain would be to use a uuid column type on postgresql (16 bytes vs 36 for varchar) but i noticed that with amp analytics in use that is not possible as the visit token sent is not conforming to a uuid format, i get values like  amp-EWydgmHg4UwiIiioxaE00Q
--

Thank you for accepting my pull request