@ankane 2021-04-24
Thanks for checking.

Ahoy tries to stick with more common data and index types by default (unless there's a significant advantage like with `jsonb` for event properties), but it's flexible enough that apps can customize as needed.